Updated formatting. Review PR notes for additional details.

+++ api/Services/permissionsService.js
@@ -1,11 +1,11 @@
 /*
-!#######################################################################
-! C-PATTM SOFTWARE
-! CRANE C-PATTM plan of action and milestones software. Use is governed by the Open Source Academic Research License Agreement contained in the file
-! crane_C_PAT.1_license.txt, which is part of this software package. BY
-! USING OR MODIFYING THIS SOFTWARE, YOU ARE AGREEING TO THE TERMS AND    
+!##########################################################################
+! CRANE PLAN OF ACTION AND MILESTONE (C-PAT) SOFTWARE
+! Use is governed by the Open Source Academic Research License Agreement
+! contained in the LICENSE.MD file, which is part of this software package.
+! BY USING OR MODIFYING THIS SOFTWARE, YOU ARE AGREEING TO THE TERMS AND    
 ! CONDITIONS OF THE LICENSE.  
-!########################################################################
+!##########################################################################
 */
 
 'use strict';
@@ -81,20 +81,7 @@ exports.postPermission = async function postPermission(userId, elevate, req) {
 
     try {
         return await withConnection(async (connection) => {
-            if (elevate) {
-                const userSql = "SELECT * FROM user WHERE userId = ?";
-                const [userRows] = await connection.query(userSql, [userId]);
-
-                if (userRows.length === 0) {
-                    throw new SmError.PrivilegeError('User not found');
-                }
-
-                const isAdmin = userRows[0].isAdmin;
-
-                if (isAdmin !== 1) {
-                    throw new SmError.PrivilegeError('User requesting Elevate without admin permissions.');
-                }
-
+            if (elevate && req.userObject.isAdmin === true) {
                 let sql_query = "INSERT INTO cpat.collectionpermissions (accessLevel, userId, collectionId) VALUES (?, ?, ?);";
                 await connection.query(sql_query, [req.body.accessLevel, req.body.userId, req.body.collectionId]);
 
@@ -154,20 +141,7 @@ exports.putPermission = async function putPermission(userId, elevate, req) {
 
     try {
         return await withConnection(async (connection) => {
-            if (elevate) {
-                const userSql = "SELECT * FROM user WHERE userId = ?";
-                const [userRows] = await connection.query(userSql, [userId]);
-
-                if (userRows.length === 0) {
-                    throw new SmError.PrivilegeError('User not found');
-                }
-
-                const isAdmin = userRows[0].isAdmin;
-
-                if (isAdmin !== 1) {
-                    throw new SmError.PrivilegeError('User requesting Elevate without admin permissions.');
-                }
-
+            if (elevate && req.userObject.isAdmin === true) {
                 let sql_query = "UPDATE cpat.collectionpermissions SET collectionId = ?, accessLevel = ? WHERE userId = ? AND collectionId = ?;";
                 await connection.query(sql_query, [req.body.newCollectionId, req.body.accessLevel, req.body.userId, req.body.oldCollectionId]);
 
@@ -209,20 +183,7 @@ exports.deletePermission = async function deletePermission(userId, elevate, req)
 
     try {
         return await withConnection(async (connection) => {
-            if (elevate) {
-                const userSql = "SELECT * FROM user WHERE userId = ?";
-                const [userRows] = await connection.query(userSql, [userId]);
-
-                if (userRows.length === 0) {
-                    throw new SmError.PrivilegeError('User not found');
-                }
-
-                const isAdmin = userRows[0].isAdmin;
-
-                if (isAdmin !== 1) {
-                    throw new SmError.PrivilegeError('User requesting Elevate without admin permissions.');
-                }
-
+            if (elevate && req.userObject.isAdmin === true) {
                 let sql = "DELETE FROM  cpat.collectionpermissions WHERE userId = ? AND collectionId = ?";
                 await connection.query(sql, [req.params.userId, req.params.collectionId]);
